Adjust startKey on list start
On list if revision != 0 and startKey != "" it is assumed
that startKey exists.  If it doesn't exists nothing will be returned.
If we are starting a new list at a revision then startKey will equal
prefix and that key does not exists.  This change will set startKey to
"" in this situaion so that list works.