Simplify ttlEvents ListWatch
There is no reason to do the list in one goroutine and the watch in another, as the watcher just blocks on startup waiting for the list to finish and send the revision on the channel. This can be simplified to one goroutine, which lets us drop the channel and waitgroup entirely.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>