Reject invalid endpoint (#320)
* Reject invalid endpoint

Prior, a non-empty endpoint of the incorrect format would cause an
embedded sqlite backend to be used. By passing a non-empty endpoint
the user is intending to use an external backend. Now, the endpoint
will be validated and passing a non-empty endpoint of an incorrect
format will result in an error.

Signed-off-by: Ricardo Weir <ricardo.weir@loft.sh>

* Update sqlite test

Signed-off-by: Ricardo Weir <ricardo.weir@loft.sh>

---------

Signed-off-by: Ricardo Weir <ricardo.weir@loft.sh>