Add support for apiserver-managed compact
This moves the outer loop from `compactor` to a new `compactIter`
function that can be called as the Compact implementation. This
replaces the current code path, which just directly compacts to
the given revision without any safeties or batching.

With safety on the Compact implementation, we can allow the apiserver
to handle compacting, as long as automatic compact is disabled.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>