Fix continuation token support
The apiserver may create "pregenerated continue tokens", where the start
key does not end with a trailing null. When this is done, the start key
should be included in the results, whereas normally it should not. The
apiserver indicates that the start key should NOT be included by
appending the trailing null, that's what it's for.

Fix this by not stripping the trailing null byte from the key and selecting rows with name >= key, not > key

Fixes continuation token tests:
* TestList/test_List_with_pregenerated_continue_token
* TestList/ignores_resource_version_0_for_List_with_pregenerated_continue_token

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>