Fix issues with TTL workqueue shutdown
The for loop was only checking for context cancellation once, before
going into an endless range on the event channel. Move the channel read
into the select, so that the for loop actually runs more than once and
checks for cancellation every iteration.

We should also shut down the workqueue if the event channel is closed,
as this is another indicator of some problem that should halt
processing.

On the processing side, don't reenqueue on context cancellation, as it
will never succeed.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>